Move `TestingFileStorage` to a separate module
This class is also intended to be used of users of this library.

So, it can't be moved to the tests directory, and also it cannot
be renamed easily.

modified:   src/flask_uploads/__init__.py
modified:   src/flask_uploads/flask_uploads.py
new file:   src/flask_uploads/test_helper.py